Reorganizada la estructura de fragmentos. Se crea el primer recycler de movimientos financieros con su respectivo cardview y su implementación utilizando Groupie
