+++ app/src/main/java/com/example/piggybank/Data/Movement.kt
@@ -1,0 +1,8 @@
+package com.example.piggybank.Data
+
+data class Movement (
+    val movementType: String,
+    val movementAmount: String,
+    val movementConcept: String
+)
+
