+++ app/src/main/java/com/example/piggybank/UI/Balance/Fragments/BalanceFragment.kt
@@ -1,0 +1,48 @@
+package com.example.piggybank.UI.Balance.Fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.piggybank.R
+import kotlinx.android.synthetic.main.safe_fragment.*;
+
+class BalanceFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.balance_fragment, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        addMoneyToSafeButton?.setOnClickListener(View.OnClickListener {
+
+        })
+
+        withdrawMoneyFromSafeButton.setOnClickListener(View.OnClickListener {
+
+        })
+
+        safeHistoryButton.setOnClickListener(View.OnClickListener {
+
+        })
+
+    }
+
+    companion object {
+        @JvmStatic
+        fun newInstance(): BalanceFragment {
+            val fragment = BalanceFragment()
+            val args = Bundle()
+            fragment.arguments = args
+            return fragment
+        }
+    }
+}
