--- app/src/main/java/com/example/piggybank/UI/Fragments/MainFragments/LoginFragment.kt
@@ -1,40 +1,0 @@
-package com.example.piggybank.UI.Fragments.MainFragments
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-import com.example.piggybank.R
-import kotlinx.android.synthetic.main.login_fragment.*
-
-class LoginFragment : Fragment() {
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.login_fragment, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        loginButton?.setOnClickListener(View.OnClickListener {
-           findNavController().navigate(R.id.login_to_menu)
-
-        })
-
-        toRegisterButton?.setOnClickListener(View.OnClickListener {
-            findNavController().navigate(R.id.login_to_registration)
-        })
-    }
-
-    companion object {
-        fun newInstance(): LoginFragment {
-            val fragment = LoginFragment()
-            val args = Bundle()
-            fragment.arguments = args
-            return fragment
-        }
-    }
-}
