+++ app/src/main/java/com/example/piggybank/UI/Safe/Fragments/WithMoneyFromSafeFragment.kt
@@ -1,0 +1,41 @@
+package com.example.piggybank.UI.Safe.Fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
+import com.example.piggybank.R
+import kotlinx.android.synthetic.main.with_money_from_safe_fragment.*
+
+class WithMoneyFromSafeFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.with_money_from_safe_fragment, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+       withMoneyFromSafeButton?.setOnClickListener(View.OnClickListener {
+           findNavController().navigate(R.id.with_to_safe)
+       })
+
+    }
+
+    companion object {
+        @JvmStatic
+        fun newInstance(): WithMoneyFromSafeFragment {
+            val fragment = WithMoneyFromSafeFragment()
+            val args = Bundle()
+            fragment.arguments = args
+            return fragment
+        }
+    }
+}
