+++ app/src/main/res/values/dimen.xml
@@ -2,6 +2,7 @@
 <resources>
     <dimen name="editText_width">256dp</dimen>
     <dimen name="normal_margin">16dp</dimen>
+    <dimen name="small_margin">8dp</dimen>
     <dimen name="vertical_elements_margin">56dp</dimen>
     <dimen name="vertical_menu_elements_margin">40dp</dimen>
     <dimen name="title_font_size">30sp</dimen>
@@ -10,4 +11,5 @@
     <dimen name="buttons_width">200dp</dimen>
     <dimen name="buttons_height">40dp</dimen>
     <dimen name="extended_buttons_height">60dp</dimen>
+    <dimen name="card_elevation">4dp</dimen>
 </resources>
